UPD & ADD converting existing JavaNamespaceConverter to interface + trait
Although I like the functionality and convenience with being able to
pass dot-separated class names I did not like having to keep up with
instantiating and passing the object around. With this implementation
if objects should be responsible for converting Java namespaces it is
an explicit part of their interface signature and we provide an easy
way to satisfy that interface. This implementation is actually more
expressive and easier to use than simple Dependency Injection